NGram_Optimizer_App_vU/ngram_optimizer_app/requirements.txt
streamlit
pandas
plotly
python-pptx
xlsxwriter
pyinstaller